Rename test bizsocket-rx to bizsocket-base
